Created sadcd class and moved chunk daemon into main.cpp file
