+++ src/sadcd/sadcd.cpp
@@ -1,56 +1,33 @@
-/* This creates the binary for the chunk server daemon, which should only be
- * started by the sadcd-bootstrap process. It opens up a listener on an ip and
- * port specified by the command line to handle incoming requests.
- */
+/* Code for the sadcd class */
 
 // sadfs-specific includes
 #include <sadfs/comm/inet.hpp>
+#include <sadfs/sadcd/sadcd.hpp>
 
 // standard includes
 #include <array>
 #include <cstring>  // std::strerror
 #include <iostream>
+#include <string>
 #include <unistd.h> // read/write
 
-namespace
+void sadcd::
+start()
 {
+	auto listener = sadfs::inet::listener{ip, port};
 
-// struct to hold info needed to set up a chunk server
-struct settings
-{
-	sadfs::inet::port_no port;
-	sadfs::inet::ip_addr ip;
-};
-
-// returns settings based on command-line arguments
-settings
-read_command_line(int argc, char *argv[])
-{
-	auto port = 0;
-	auto ip = "";
-	for (auto i = 1; i < argc - 1; i ++)
+	while (true)
 	{
-		auto arg = std::string{argv[i]};
-		if (arg.compare("--port") == 0)
-		{
-			i++;
-			port = std::stoi(std::string{argv[i]});
-		}
-		else if (arg.compare("--ipaddress") == 0)
-		{
-			i++;
-			ip = argv[i];
-		}
-	}
+		auto sock = listener.accept();
+		auto result = process_message(sock);
 
-	// TODO: verify that port fits in 16 bits in edge case that this program is
-	// started by some method other than sadcd-bootstrap
-	// should also double check that an ip and port number were in fact passed
-	return settings{static_cast<std::uint16_t>(port), ip};
+		std::cout << result << "\n";
+		// perform some action based on result
+	}
 }
 
 // reads the message from a socket that just received some data
-std::string
+std::string sadcd::
 process_message(sadfs::socket& sock)
 {
 	auto buf = std::array<char, 512>{};
@@ -77,26 +54,4 @@ process_message(sadfs::socket& sock)
 	}
 
 	return result;
-}
-
-} // unnamed namespace
-
-int 
-main(int argc, char *argv[])
-{
-
-	// populate server_settings from command line args
-	auto server_settings = read_command_line(argc, argv);
-
-	// set up a listener for incoming messages
-	auto listener = sadfs::inet::listener{server_settings.ip, server_settings.port};
-
-	while (true)
-	{
-		auto sock = listener.accept();
-		auto result = process_message(sock);
-
-		std::cout << result << "\n";
-		// perform some action based on result
-	}
 }
